requirements.txt
alembic==1.4.0
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dateparser==0.7.4
dialogflow==1.0.0
dialogflow-fulfillment==0.1.2
Flask==1.1.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.2
fuzzywuzzy==0.18.0
google-api-core==1.17.0
google-auth==1.15.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydialogflow-fulfillment==0.1.3
pydocstyle==5.0.2
pyflakes==2.2.0
pylama==7.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-Levenshtein==0.12.0
pytz==2019.3
regex==2020.5.14
requests==2.23.0
rsa==4.0
schedule==0.6.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.17
state-machine==0.2.10
tzlocal==2.1
urllib3==1.25.8
vedis==0.7.1
vk-api==11.8.0
Werkzeug==1.0.0
alembic==1.4.0
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Cython==0.29.19
dateparser==0.7.4
dialogflow==1.0.0
dialogflow-fulfillment==0.1.2
Flask==1.1.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.2
fuzzywuzzy==0.18.0
google-api-core==1.17.0
google-auth==1.15.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydialogflow-fulfillment==0.1.3
pydocstyle==5.0.2
pyflakes==2.2.0
pylama==7.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-Levenshtein==0.12.0
pytz==2019.3
regex==2020.5.14
requests==2.23.0
rsa==4.0
schedule==0.6.0
singleton-decorator==1.0.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.17
telegram==0.0.1
tzlocal==2.1
urllib3==1.25.8
vk-api==11.8.0
Werkzeug==1.0.0
alembic==1.4.0
backcall==0.1.0
bitlyshortener==0.3.5
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Cython==0.29.19
dateparser==0.7.4
decorator==4.4.2
dialogflow==1.0.0
dialogflow-fulfillment==0.1.2
Flask==1.1.2
Flask-Migrate==2.5.2
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.2
fuzzywuzzy==0.18.0
google-api-core==1.17.0
google-auth==1.15.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydialogflow-fulfillment==0.1.3
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pylama==7.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-Levenshtein==0.12.0
python-state-machine==0.0.18
pytz==2019.3
regex==2020.5.14
requests==2.23.0
rsa==4.0
schedule==0.6.0
singleton-decorator==1.0.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.17
state-machine==0.2.10
telegram==0.0.1
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.8
vedis==0.7.1
vk-api==11.8.0
wcwidth==0.2.2
Werkzeug==1.0.0
